Implement Foreach loop in store page and finish implementation of database factory & seeder

+++ database/migrations/2024_10_14_075444_create_products_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->integer('price');
             $table->string('image_path');
             $table->double('star_rating');
-            $table->double('count_rating');
+            $table->integer('count_rating');
             $table->foreignId('user_id')->constrained( //seller
                 table: 'users',
                 indexName: 'products_user_id'
